lint: removes unused imports

+++ test/functional/wallet_import_rescan.py
@@ -25,7 +25,6 @@ from test_framework.address import (
     AddressType,
     ADDRESS_BCRT1_UNSPENDABLE,
 )
-from test_framework.messages import COIN
 from test_framework.util import (
     assert_equal,
     set_node_times,
@@ -278,7 +277,7 @@ class ImportRescanTest(BitcoinTestFramework):
             variant.key = self.nodes[1].dumpprivkey(variant.address["address"])
             # Ensure output is large enough to pay for fees: conservatively assuming txsize of
             # 500 vbytes and feerate of 20 sats/vbytes
-            variant.initial_amount = get_rand_amount(min_amount=((sat_to_btc(500 * 20)) + AMOUNT_DUST))
+            variant.initial_amount = get_rand_amount(min_amount=float(sat_to_btc(500 * 20) + Decimal(AMOUNT_DUST)))
             variant.initial_txid = self.nodes[0].sendtoaddress(variant.address["address"], variant.initial_amount)
             variant.confirmation_height = 0
             variant.timestamp = timestamp
